+++ util/copyUrl.ts
@@ -1,0 +1,5 @@
+export const handleCopyUrl = () => {
+  const currentUrl = window.location.href;
+  navigator.clipboard.writeText(currentUrl);
+  //토스트 연결 필요
+};
